@@ -16,9 +16,9 @@
 
 package com.benoitletondor.mvp.core.fragment;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.annotation.VisibleForTesting;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.annotation.VisibleForTesting;
 
 import com.benoitletondor.mvp.core.SpyPresenter;
 import com.benoitletondor.mvp.core.presenter.loader.PresenterFactory;
