@@ -17,9 +17,9 @@
 package com.benoitletondor.mvp.core.presenter.loader;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.content.Loader;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.loader.content.Loader;
 
 import com.benoitletondor.mvp.core.presenter.Presenter;
 
