@@ -1,7 +1,7 @@
 package com.benoitletondor.mvp.core.sample.injection;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.benoitletondor.mvp.core.sample.App;
 
